Refactored classes, implemented hashcode() and equals()

+++ src/main/java/by/jenka/jpoker/suit/Suit.java
@@ -3,5 +3,5 @@ package by.jenka.jpoker.suit;
 import by.jenka.jpoker.common.Unicodable;
 
 public interface Suit extends Unicodable {
-    int getValue();
+    long getValue();
 }
